+++ src/modules/sbstudio/plugin/utils/sampling.py
@@ -1,14 +1,63 @@
 from bpy.types import Context, Object
 from collections import defaultdict
 from itertools import chain
-from typing import Dict, Iterable, Optional, Sequence, Tuple
+from typing import Dict, Iterable, Iterator, Optional, Sequence, Tuple
 
+from sbstudio.model.color import Color4D
+from sbstudio.model.light_program import LightProgram
 from sbstudio.model.point import Point4D
 from sbstudio.model.trajectory import Trajectory
+from sbstudio.plugin.materials import get_led_light_color
+from sbstudio.plugin.utils.evaluator import get_position_of_object
 
 from .decorators import with_context
 
-__all__ = ("sample_positions_of_objects", "sample_positions_of_objects_in_frame_range")
+__all__ = (
+    "each_frame_in",
+    "frame_range",
+    "sample_colors_of_objects",
+    "sample_positions_of_objects",
+    "sample_positions_of_objects_in_frame_range",
+    "sample_positions_and_colors_of_objects",
+)
+
+
+def _to_int_255(value: float) -> int:
+    """Convert [0,1] float to clamped [0,255] int."""
+    return int(max(0, min(255, round(value * 255))))
+
+
+@with_context
+def frame_range(
+    start: int, end: int, *, fps: int, context: Optional[Context] = None
+) -> Iterator[int]:
+    """Generator that iterates from the given start frame to the given end frame
+    with the given number of frames per second.
+    """
+    assert context is not None  # injected
+
+    scene_fps = context.scene.render.fps
+    frame_step = max(1, int(scene_fps // fps))
+    return chain(range(start, end, frame_step), [end])
+
+
+@with_context
+def each_frame_in(
+    frames: Iterable[int], *, context: Optional[Context] = None
+) -> Iterable[Tuple[int, float]]:
+    """Generator that iterates over the given frames, sets the current frame in
+    Blender to each frame one by one, and then yields a tuple consisting of
+    the current frame _index_ and the current frame _timestamp_ to the caller.
+    """
+    assert context is not None  # injected
+
+    scene = context.scene
+    fps = scene.render.fps
+
+    for frame in frames:
+        scene.frame_set(frame)
+        time = frame / fps
+        yield frame, time
 
 
 @with_context
@@ -17,6 +66,7 @@ def sample_positions_of_objects(
     frames: Iterable[int],
     *,
     by_name: bool = False,
+    simplify: bool = False,
     context: Optional[Context] = None,
 ) -> Dict[Object, Trajectory]:
     """Samples the positions of the given Blender objects at the given frames,
@@ -29,31 +79,140 @@ def sample_positions_of_objects(
             of the objects
         context: the Blender execution context; `None` means the current
             Blender context
+        simplify: whether to simplify the trajectories. If this option is
+            enabled, the resulting trajectories might not contain samples for
+            all the input frames; excess samples that are identical to previous
+            ones will be removed.
 
     Returns:
         a dictionary mapping the objects to their trajectories
     """
     trajectories = defaultdict(Trajectory)
-    fps = context.scene.render.fps
 
-    for frame in frames:
-        context.scene.frame_set(frame)
+    for _, time in each_frame_in(frames, context=context):
         for obj in objects:
             key = obj.name if by_name else obj
-            trajectories[key].append(
-                Point4D(frame / fps, *(obj.matrix_world.translation))
+            trajectories[key].append(Point4D(time, *(obj.matrix_world.translation)))
+
+    if simplify:
+        return {key: value.simplify_in_place() for key, value in trajectories.items()}
+    else:
+        return dict(trajectories)
+
+
+@with_context
+def sample_colors_of_objects(
+    objects: Sequence[Object],
+    frames: Iterable[int],
+    *,
+    by_name: bool = False,
+    simplify: bool = False,
+    context: Optional[Context] = None,
+) -> Dict[Object, LightProgram]:
+    """Samples the colors of the given Blender objects at the given frames,
+    returning a dictionary mapping the objects to their light programs.
+
+    Parameters:
+        objects: the Blender objects to process
+        frames: an iterable yielding the indices of the frames to process
+        by_name: whether the result dictionary should be keyed by the _names_
+            of the objects
+        simplify: whether to simplify the light programs. If this option is
+            enabled, the resulting light program might not contain samples
+            for all the input frames; excess samples that are identical to
+            previous ones will be removed.
+        context: the Blender execution context; `None` means the current
+            Blender context
+
+    Returns:
+        a dictionary mapping the objects to their light programs
+    """
+    lights = defaultdict(LightProgram)
+
+    for _, time in each_frame_in(frames, context=context):
+        for obj in objects:
+            key = obj.name if by_name else obj
+            color = get_led_light_color(obj)
+            lights[key].append(
+                Color4D(
+                    time,
+                    _to_int_255(color[0]),
+                    _to_int_255(color[1]),
+                    _to_int_255(color[2]),
+                )
             )
 
-    return dict(trajectories)
+    if simplify:
+        return {key: value.simplify() for key, value in lights.items()}
+    else:
+        return dict(lights)
+
+
+@with_context
+def sample_positions_and_colors_of_objects(
+    objects: Sequence[Object],
+    frames: Iterable[int],
+    *,
+    by_name: bool = False,
+    simplify: bool = False,
+    context: Optional[Context] = None,
+) -> Dict[Object, Tuple[Trajectory, LightProgram]]:
+    """Samples the positions and colors of the given Blender objects at the
+    given frames, returning a dictionary mapping the objects to their
+    trajectories and light programs.
+
+    Parameters:
+        objects: the Blender objects to process
+        frames: an iterable yielding the indices of the frames to process
+        by_name: whether the result dictionary should be keyed by the _names_
+            of the objects
+        simplify: whether to simplify the trajectories and light programs. If
+            this option is enabled, the resulting trajectories and light programs
+            might not contain samples for all the input frames; excess samples
+            that are identical to previous ones will be removed.
+        context: the Blender execution context; `None` means the current
+            Blender context
+
+    Returns:
+        a dictionary mapping the objects to their trajectories and light programs
+    """
+    trajectories = defaultdict(Trajectory)
+    lights = defaultdict(LightProgram)
+
+    for _, time in each_frame_in(frames, context=context):
+        for obj in objects:
+            key = obj.name if by_name else obj
+            pos = get_position_of_object(obj)
+            color = get_led_light_color(obj)
+            trajectories[key].append(Point4D(time, *pos))
+            lights[key].append(
+                Color4D(
+                    time,
+                    _to_int_255(color[0]),
+                    _to_int_255(color[1]),
+                    _to_int_255(color[2]),
+                )
+            )
+
+    if simplify:
+        return {
+            key: (trajectory.simplify_in_place(), lights[key].simplify())
+            for key, trajectory in trajectories.items()
+        }
+    else:
+        return {
+            key: (trajectory, lights[key]) for key, trajectory in trajectories.items()
+        }
 
 
 @with_context
 def sample_positions_of_objects_in_frame_range(
     objects: Sequence[Object],
-    frame_range: Tuple[int, int],
+    bounds: Tuple[int, int],
     *,
     fps: int,
     by_name: bool = False,
+    simplify: bool = False,
     context: Optional[Context] = None,
 ) -> Dict[Object, Trajectory]:
     """Samples the positions of the given Blender objects in the given range
@@ -62,21 +221,26 @@ def sample_positions_of_objects_in_frame_range(
 
     Parameters:
         objects: the Blender objects to process
-        frame_range: the range of frames to process; both ends are inclusive
+        bounds: the range of frames to process; both ends are inclusive
         fps: the desired number of frames per second; the result may contain
             more samples per frame if the scene FPS is not an exact multiple
             of the desired FPS
         by_name: whether the result dictionary should be keyed by the _names_
             of the objects
+        simplify: whether to simplify the trajectories. If this option is
+            enabled, the resulting trajectories might not contain samples for
+            all the input frames; excess samples that are identical to previous
+            ones will be removed.
         context: the Blender execution context; `None` means the current
             Blender context
 
     Returns:
         a dictionary mapping the objects (or their names) to their trajectories
     """
-    scene_fps = context.scene.render.fps
-    frame_step = max(1, int(scene_fps // fps))
-    frames = chain(range(frame_range[0], frame_range[1], frame_step), [frame_range[1]])
     return sample_positions_of_objects(
-        objects, frames, by_name=by_name, context=context
+        objects,
+        frame_range(bounds[0], bounds[1], fps=fps, context=context),
+        by_name=by_name,
+        simplify=simplify,
+        context=context,
     )